rename Tween to Interpolator
In most (if not all) tweening libraries, the Tween is not only responsible for interpolating and updating values, but also for updating itself.
Currently, all Tween really does is the former, so let's give it a more specific name, and see if we can create a Tween thats more useful for the common usecase of animating browser objects.